Continue develop bufr2nc convertor. Add kind and time module.

+++ src/prepbufr-fortran/kinds.f90
@@ -42,6 +42,6 @@ module kinds
 
 !
   real(r_single),parameter,public :: rmissing=-99999.0
-  real(i_kind),parameter,public :: imissing=-99999
+  integer(i_kind),parameter,public :: imissing=-99999
 
 end module kinds
